I finish View Model for App and MVVM
