+++ app/src/main/java/com/example/newsappinkotlin/network/NewsClient.kt
@@ -17,32 +17,7 @@ object NewsClient {
         service =retrofit.create(APIServices::class.java)
     }
     // to be added to the view model
-    fun fetchNews(page:Int=1, onSuccess:(newsList:MutableList<NewsModel>)->Unit, onError:()->Unit)
-    {
-        service.getAllNews(pageNumber = page).enqueue(object : Callback<NewsResponse>
-        {
-            override fun onFailure(call: Call<NewsResponse>, t: Throwable) {
-                onError.invoke()
-            }
 
-            override fun onResponse(
-                call: Call<NewsResponse>,
-                response: Response<NewsResponse>
-            ) {
-                if (response.isSuccessful)
-                {
-                    if (response.body()!=null)
-                    {
-                        onSuccess.invoke(response.body()!!.articles)
-                    }
-                    else
-                    {
-                        onError.invoke()
-                    }
-                }
-            }
-        })
-    }
 // to be added to the view model
     fun fetchHeadlines(page:Int=1, onSuccess:(newsList:MutableList<NewsModel>)->Unit, onError:()->Unit)
     {
