+++ app/src/main/java/com/example/newsappinkotlin/ui/ViewModel/NewsViewModel.kt
@@ -1,59 +1,37 @@
 package com.example.newsappinkotlin.ui.ViewModel
 
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.example.newsappinkotlin.model.NewsModel
+import com.example.newsappinkotlin.network.NewsClient
+import com.example.newsappinkotlin.network.NewsResponse
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
-class NewsViewModel {
-    var NewsSourceMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsAuthorMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsTitleMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsDesciptionMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsUrlMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsUrltoimageMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsPublishedatMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
-    var NewsContentMutableLiveData: MutableLiveData<String> = MutableLiveData<String>()
+class NewsViewModel : ViewModel() {
+    var mutableNewsList : MutableLiveData<ArrayList<NewsModel>> = MutableLiveData()
+    fun fetchNews(page:Int=1 )
+    {
+        NewsClient.service.getAllNews(pageNumber = page).enqueue(object : Callback<NewsResponse>
+        {
+            override fun onFailure(call: Call<NewsResponse>, t: Throwable) {
+ //               onError.invoke()
+            }
 
-    private fun getNews(): NewsModel = NewsModel(
-        source = "Al-Ahram",
-        author = "Frouk",
-        title = "EgyptAir to resume flights to five destinations starting 1 October",
-        desciption = "Egypt will resume direct flights to.............",
-        url = "http://english.ahram.org.eg",
-        urlToImage = "http://english.ahram.org.eg",
-        publishedAt = "27-9-2020",
-        content = "Egypt will resume direct flights to Oman, Jordan, Rwanda,"
-    )
+            override fun onResponse(
+                call: Call<NewsResponse>,
+                response: Response<NewsResponse>
+            ) {
+                if (response.isSuccessful)
+                {
+                    if (response.body()!=null)
+                    {
+                        mutableNewsList.value = response?.body()!!.articles as ArrayList<NewsModel>
+                    }
 
-    fun getNewsSource() {
-        var data = getNews().source
-        NewsSourceMutableLiveData.value = data
-    }
-    fun getNewsAuthor() {
-        var data = getNews().source
-        NewsAuthorMutableLiveData.value = data
-    }
-    fun getNewsTitle() {
-        var data = getNews().title
-        NewsTitleMutableLiveData.value = data
-    }
-    fun getNewsDesciption() {
-        var data = getNews().source
-        NewsDesciptionMutableLiveData.value = data
-    }
-    fun getNewsUrl() {
-        var data = getNews().source
-        NewsUrlMutableLiveData.value = data
-    }
-    fun getNewsUrltoimage() {
-        var data = getNews().source
-        NewsUrltoimageMutableLiveData.value = data
-    }
-    fun getNewspublishedat() {
-        var data = getNews().source
-        NewsPublishedatMutableLiveData.value = data
-    }
-    fun getNewsContent() {
-        var data = getNews().source
-        NewsContentMutableLiveData.value = data
+                }
+            }
+        })
     }
 }
